requirements.txt
fastapi==0.65.2
uvicorn==0.11.7
pydantic
# pydantic==1.6.2
starlette==0.14.2
python-multipart==0.0.5
requests==2.24.0
# scikit-learn==1.0.1
scikit-learn==1.0.2
joblib==0.16.0
gunicorn==20.0.4
pytest==6.2.2
locust==1.4.3
codecov==2.1.13
pytest-cov==2.11.1
pandas-profiling==3.6.6
